% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive-cols.R
\name{as_reactive_logical}
\alias{as_reactive_logical}
\title{Typed columns that contain a reactive expression}
\usage{
as_reactive_logical(x, expr)
}
\description{
Typed columns that contain a reactive expression
}
