% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\title{A mutable tibble/data.frame to enable a forward/reverse graphics pipeline}
\usage{
pin(.data)
}
\description{
Here we simply just store the data in an envirionment then all
`dplyr` verbs just act on the data frame inside the environment, making
the data.frame/tibble inside a mutable object. We wrap this environment
into an s3 class called a 'mutibble'.
}
\details{
We need to make it obvious that we are using a mutable data structure,
so we use the verb `pin` to emphasise that we are 'pinning' down the input
data to an environment.
}
