% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{put_values}
\alias{put_values}
\title{Mutate in place}
\usage{
put_values(.data, vals)
}
\description{
The pin allows a tibble to become mutable, and in order to update values
inside of this new mutable tibble, we need to 'put' values in there.
}
\details{
put is essentially a call to mutate but is based on the condition that a
event has taken place (hence it uses dplyr's case_when function)
we need to think of a way of providing sensible defaults with the possible
things that could be modified here (i.e. colour or shape or other features)
could aesthetic variables be placed here (maybe with the aes_ modifier?)?
}
