% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-mutate.R, R/summarise.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mutate}
\alias{transmute}
\alias{summarise}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{summarise}}}
}}

