% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-shiny.R
\name{render_shiny}
\alias{render_shiny}
\title{just a brush
vis <- mtcars %>% visualise(x = hp, y = mpg) %>% draw_points()
vis_b <- vis %>% control_drag("brush") %>% draw_rect()
transient brush with highlight
vis_t <- vis %>% mutate(
                         sel = inside(aes_x, aes_y, vis_b$event),
                         aes_fill = ifelse(sel, "black", "white")
                        ) %>%
                        mesh(vis_b)
persisent selection brush
vis_p <- vis %>% 
               mutate(sel = FALSE) %>% 
               mutate_persistent(sel = sel | inside(aes_x, aes_y, vis_b$event)) %>%
               mesh(vis_b)}
\usage{
render_shiny(x)
}
\description{
just a brush
vis <- mtcars %>% visualise(x = hp, y = mpg) %>% draw_points()
vis_b <- vis %>% control_drag("brush") %>% draw_rect()
transient brush with highlight
vis_t <- vis %>% mutate(
                         sel = inside(aes_x, aes_y, vis_b$event),
                         aes_fill = ifelse(sel, "black", "white")
                        ) %>%
                        mesh(vis_b)
persisent selection brush
vis_p <- vis %>% 
               mutate(sel = FALSE) %>% 
               mutate_persistent(sel = sel | inside(aes_x, aes_y, vis_b$event)) %>%
               mesh(vis_b)
}
