% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_lm}
\alias{summarise_lm}
\title{getting fitted values from a model is more of a transform?}
\usage{
summarise_lm(.tbl_pl, formula = y ~ x, ...)
}
\description{
getting fitted values from a model is more of a transform?
}
