% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{set_opts}
\alias{set_opts}
\title{The API for drawing is quite simple:
It takes a plot tibble and appends a graphical primitive + 
any additional options to to the plot, layers are then built
up using mesh.}
\usage{
set_opts(.data, opts, geom)
}
\description{
How to handle when options that overlap aesthetics?  
 1. Throw a warning/message?
 2. overwrite it?
}
